refactor: rename gstack binaries and directories to nstack
Rename all bin/gstack-* scripts to bin/nstack-*, update internal
references (env vars, cross-binary calls, storage paths). Rename
gstack-upgrade/ to nstack-upgrade/ and open-gstack-browser/ to
open-nstack-browser/. Rename scripts/app/gstack-browser to
nstack-browser.
