+++ bin/dev-setup
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
-# Set up gstack for local development — test skills from within this repo.
+# Set up nstack for local development — test skills from within this repo.
 #
-# Creates .claude/skills/gstack → (symlink to repo root) so Claude Code
+# Creates .claude/skills/nstack → (symlink to repo root) so Claude Code
 # discovers skills from your working tree. Changes take effect immediately.
 #
 # Also copies .env from the main worktree if this is a Conductor workspace
@@ -12,6 +12,17 @@
 set -e
 
 REPO_ROOT="$(cd "$(dirname "$0")/.." && pwd)"
+
+# 0. Ensure bun is on PATH (version managers / Conductor may not source shell profile)
+export BUN_INSTALL="${BUN_INSTALL:-$HOME/.bun}"
+if ! command -v bun &>/dev/null && [ -x "$BUN_INSTALL/bin/bun" ]; then
+  export PATH="$BUN_INSTALL/bin:$PATH"
+fi
+if ! command -v bun &>/dev/null; then
+  echo "Error: bun is not installed." >&2
+  echo "Install it with: curl -fsSL https://bun.sh/install | bash" >&2
+  exit 1
+fi
 
 # 1. Copy .env from main worktree (if we're a worktree and don't have one)
 if [ ! -f "$REPO_ROOT/.env" ]; then
@@ -31,38 +42,38 @@ fi
 # 3. Create .claude/skills/ inside the repo
 mkdir -p "$REPO_ROOT/.claude/skills"
 
-# 4. Symlink .claude/skills/gstack → repo root
+# 4. Symlink .claude/skills/nstack → repo root
 # This makes setup think it's inside a real .claude/skills/ directory
-GSTACK_LINK="$REPO_ROOT/.claude/skills/gstack"
-if [ -L "$GSTACK_LINK" ]; then
+NSTACK_LINK="$REPO_ROOT/.claude/skills/nstack"
+if [ -L "$NSTACK_LINK" ]; then
   echo "Updating existing symlink..."
-  rm "$GSTACK_LINK"
-elif [ -d "$GSTACK_LINK" ]; then
-  echo "Error: .claude/skills/gstack is a real directory, not a symlink." >&2
+  rm "$NSTACK_LINK"
+elif [ -d "$NSTACK_LINK" ]; then
+  echo "Error: .claude/skills/nstack is a real directory, not a symlink." >&2
   echo "Remove it manually if you want to use dev mode." >&2
   exit 1
 fi
-ln -s "$REPO_ROOT" "$GSTACK_LINK"
+ln -s "$REPO_ROOT" "$NSTACK_LINK"
 
-# 5. Create .agents/skills/gstack → repo root (for Codex/Gemini/Cursor)
+# 5. Create .agents/skills/nstack → repo root (for Codex/Gemini/Cursor)
 mkdir -p "$REPO_ROOT/.agents/skills"
-AGENTS_LINK="$REPO_ROOT/.agents/skills/gstack"
+AGENTS_LINK="$REPO_ROOT/.agents/skills/nstack"
 if [ -L "$AGENTS_LINK" ]; then
   rm "$AGENTS_LINK"
 elif [ -d "$AGENTS_LINK" ]; then
-  echo "Warning: .agents/skills/gstack is a real directory, skipping." >&2
+  echo "Warning: .agents/skills/nstack is a real directory, skipping." >&2
 fi
 if [ ! -e "$AGENTS_LINK" ]; then
   ln -s "$REPO_ROOT" "$AGENTS_LINK"
 fi
 
 # 6. Run setup via the symlink so it detects .claude/skills/ as its parent
-"$GSTACK_LINK/setup"
+"$NSTACK_LINK/setup"
 
 echo ""
 echo "Dev mode active. Skills resolve from this working tree."
-echo "  .claude/skills/gstack → $REPO_ROOT"
-echo "  .agents/skills/gstack → $REPO_ROOT"
+echo "  .claude/skills/nstack → $REPO_ROOT"
+echo "  .agents/skills/nstack → $REPO_ROOT"
 echo "Edit any SKILL.md and test immediately — no copy/deploy needed."
 echo ""
 echo "To tear down: bin/dev-teardown"
