--- bin/gstack-config
@@ -1,93 +1,0 @@
-#!/usr/bin/env bash
-# gstack-config — read/write ~/.gstack/config.yaml
-#
-# Usage:
-#   gstack-config get <key>          — read a config value
-#   gstack-config set <key> <value>  — write a config value
-#   gstack-config list               — show all config
-#
-# Env overrides (for testing):
-#   GSTACK_STATE_DIR  — override ~/.gstack state directory
-set -euo pipefail
-
-STATE_DIR="${GSTACK_STATE_DIR:-$HOME/.gstack}"
-CONFIG_FILE="$STATE_DIR/config.yaml"
-
-# Annotated header for new config files. Written once on first `set`.
-CONFIG_HEADER='# gstack configuration — edit freely, changes take effect on next skill run.
-# Docs: https://github.com/garrytan/gstack
-#
-# ─── Behavior ────────────────────────────────────────────────────────
-# proactive: true           # Auto-invoke skills when your request matches one.
-#                           # Set to false to only run skills you type explicitly.
-#
-# routing_declined: false   # Set to true to skip the CLAUDE.md routing injection
-#                           # prompt. Set back to false to be asked again.
-#
-# ─── Telemetry ───────────────────────────────────────────────────────
-# telemetry: anonymous      # off | anonymous | community
-#                           #   off       — no data sent, no local analytics
-#                           #   anonymous — counter only, no device ID
-#                           #   community — usage data + stable device ID
-#
-# ─── Updates ─────────────────────────────────────────────────────────
-# auto_upgrade: false       # true = silently upgrade on session start
-# update_check: true        # false = suppress version check notifications
-#
-# ─── Skill naming ────────────────────────────────────────────────────
-# skill_prefix: false       # true = namespace skills as /gstack-qa, /gstack-ship
-#                           # false = short names /qa, /ship
-#
-# ─── Advanced ────────────────────────────────────────────────────────
-# codex_reviews: enabled    # disabled = skip Codex adversarial reviews in /ship
-# gstack_contributor: false # true = file field reports when gstack misbehaves
-# skip_eng_review: false    # true = skip eng review gate in /ship (not recommended)
-#
-'
-
-case "${1:-}" in
-  get)
-    KEY="${2:?Usage: gstack-config get <key>}"
-    # Validate key (alphanumeric + underscore only)
-    if ! printf '%s' "$KEY" | grep -qE '^[a-zA-Z0-9_]+$'; then
-      echo "Error: key must contain only alphanumeric characters and underscores" >&2
-      exit 1
-    fi
-    grep -E "^${KEY}:" "$CONFIG_FILE" 2>/dev/null | tail -1 | awk '{print $2}' | tr -d '[:space:]' || true
-    ;;
-  set)
-    KEY="${2:?Usage: gstack-config set <key> <value>}"
-    VALUE="${3:?Usage: gstack-config set <key> <value>}"
-    # Validate key (alphanumeric + underscore only)
-    if ! printf '%s' "$KEY" | grep -qE '^[a-zA-Z0-9_]+$'; then
-      echo "Error: key must contain only alphanumeric characters and underscores" >&2
-      exit 1
-    fi
-    mkdir -p "$STATE_DIR"
-    # Write annotated header on first creation
-    if [ ! -f "$CONFIG_FILE" ]; then
-      printf '%s' "$CONFIG_HEADER" > "$CONFIG_FILE"
-    fi
-    # Escape sed special chars in value and drop embedded newlines
-    ESC_VALUE="$(printf '%s' "$VALUE" | head -1 | sed 's/[&/\]/\\&/g')"
-    if grep -qE "^${KEY}:" "$CONFIG_FILE" 2>/dev/null; then
-      # Portable in-place edit (BSD sed uses -i '', GNU sed uses -i without arg)
-      _tmpfile="$(mktemp "${CONFIG_FILE}.XXXXXX")"
-      sed "/^${KEY}:/s/.*/${KEY}: ${ESC_VALUE}/" "$CONFIG_FILE" > "$_tmpfile" && mv "$_tmpfile" "$CONFIG_FILE"
-    else
-      echo "${KEY}: ${VALUE}" >> "$CONFIG_FILE"
-    fi
-    # Auto-relink skills when prefix setting changes (skip during setup to avoid recursive call)
-    if [ "$KEY" = "skill_prefix" ] && [ -z "${GSTACK_SETUP_RUNNING:-}" ]; then
-      GSTACK_RELINK="$(dirname "$0")/gstack-relink"
-      [ -x "$GSTACK_RELINK" ] && "$GSTACK_RELINK" || true
-    fi
-    ;;
-  list)
-    cat "$CONFIG_FILE" 2>/dev/null || true
-    ;;
-  *)
-    echo "Usage: gstack-config {get|set|list} [key] [value]"
-    exit 1
-    ;;
-esac
