--- bin/gstack-review-log
@@ -1,18 +1,0 @@
-#!/usr/bin/env bash
-# gstack-review-log — atomically log a review result
-# Usage: gstack-review-log '{"skill":"...","timestamp":"...","status":"..."}'
-set -euo pipefail
-SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
-eval "$("$SCRIPT_DIR/gstack-slug" 2>/dev/null)"
-GSTACK_HOME="${GSTACK_HOME:-$HOME/.gstack}"
-mkdir -p "$GSTACK_HOME/projects/$SLUG"
-
-# Validate: input must be parseable JSON (reject malformed or injection attempts)
-INPUT="$1"
-if ! printf '%s' "$INPUT" | bun -e "JSON.parse(await Bun.stdin.text())" 2>/dev/null; then
-  # Not valid JSON — refuse to append
-  echo "gstack-review-log: invalid JSON, skipping" >&2
-  exit 1
-fi
-
-echo "$INPUT" >> "$GSTACK_HOME/projects/$SLUG/$BRANCH-reviews.jsonl"
