--- bin/gstack-review-read
@@ -1,12 +1,0 @@
-#!/usr/bin/env bash
-# gstack-review-read — read review log and config for dashboard
-# Usage: gstack-review-read
-set -euo pipefail
-SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
-eval "$("$SCRIPT_DIR/gstack-slug" 2>/dev/null)"
-GSTACK_HOME="${GSTACK_HOME:-$HOME/.gstack}"
-cat "$GSTACK_HOME/projects/$SLUG/$BRANCH-reviews.jsonl" 2>/dev/null || echo "NO_REVIEWS"
-echo "---CONFIG---"
-"$SCRIPT_DIR/gstack-config" get skip_eng_review 2>/dev/null || echo "false"
-echo "---HEAD---"
-git rev-parse --short HEAD 2>/dev/null || echo "unknown"
