--- bin/gstack-timeline-log
@@ -1,34 +1,0 @@
-#!/usr/bin/env bash
-# gstack-timeline-log — append a timeline event to the project timeline
-# Usage: gstack-timeline-log '{"skill":"review","event":"started","branch":"main"}'
-#
-# Session timeline: local-only, never sent anywhere.
-# Required fields: skill, event (started|completed).
-# Optional: branch, outcome, duration_s, session, ts.
-# Validation failure → skip silently (non-blocking).
-set -euo pipefail
-SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
-eval "$("$SCRIPT_DIR/gstack-slug" 2>/dev/null)"
-GSTACK_HOME="${GSTACK_HOME:-$HOME/.gstack}"
-mkdir -p "$GSTACK_HOME/projects/$SLUG"
-
-INPUT="$1"
-
-# Validate: input must be parseable JSON with required fields
-if ! printf '%s' "$INPUT" | bun -e "
-  const j = JSON.parse(await Bun.stdin.text());
-  if (!j.skill || !j.event) process.exit(1);
-" 2>/dev/null; then
-  exit 0  # skip silently, non-blocking
-fi
-
-# Inject timestamp if not present
-if ! printf '%s' "$INPUT" | bun -e "const j=JSON.parse(await Bun.stdin.text()); if(!j.ts) process.exit(1)" 2>/dev/null; then
-  INPUT=$(printf '%s' "$INPUT" | bun -e "
-    const j = JSON.parse(await Bun.stdin.text());
-    j.ts = new Date().toISOString();
-    console.log(JSON.stringify(j));
-  " 2>/dev/null) || true
-fi
-
-echo "$INPUT" >> "$GSTACK_HOME/projects/$SLUG/timeline.jsonl"
