+++ bin/nstack-open-url
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+# nstack-open-url — cross-platform URL opener
+#
+# Usage: nstack-open-url <url>
+set -euo pipefail
+
+URL="${1:?Usage: nstack-open-url <url>}"
+
+case "$(uname -s)" in
+  Darwin)  open "$URL" ;;
+  Linux)   xdg-open "$URL" 2>/dev/null || echo "$URL" ;;
+  MINGW*|MSYS*|CYGWIN*) start "$URL" ;;
+  *)       echo "$URL" ;;
+esac
