+++ bin/nstack-platform-detect
@@ -1,0 +1,27 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+# nstack-platform-detect: show which AI coding agents are installed and nstack status
+# Config-driven: reads host definitions from hosts/*.ts via host-config-export.ts
+
+SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
+NSTACK_DIR="$(cd "$SCRIPT_DIR/.." && pwd)"
+
+printf "%-16s %-10s %-40s %s\n" "Agent" "Version" "Skill Path" "nstack"
+printf "%-16s %-10s %-40s %s\n" "-----" "-------" "----------" "------"
+
+for host in $(bun run "$NSTACK_DIR/scripts/host-config-export.ts" list 2>/dev/null); do
+  cmd=$(bun run "$NSTACK_DIR/scripts/host-config-export.ts" get "$host" cliCommand 2>/dev/null)
+  root=$(bun run "$NSTACK_DIR/scripts/host-config-export.ts" get "$host" globalRoot 2>/dev/null)
+  spath="$HOME/$root"
+
+  if command -v "$cmd" >/dev/null 2>&1; then
+    ver=$("$cmd" --version 2>/dev/null | head -1 || echo "unknown")
+    if [ -d "$spath" ] || [ -L "$spath" ]; then
+      status="INSTALLED"
+    else
+      status="NOT INSTALLED"
+    fi
+    printf "%-16s %-10s %-40s %s\n" "$host" "$ver" "$spath" "$status"
+  fi
+done
